feat: Added protected routes
