+++ app/(auth)/_layout.tsx
@@ -5,6 +5,7 @@ import { Stack } from "expo-router";
 const AuthLayout = () => {
   return (
     <Stack>
+      <Stack.Screen name="welcome" options={{ headerShown: false }} />
       <Stack.Screen name="login" options={{ headerShown: false }} />
       <Stack.Screen name="register" options={{ headerShown: false }} />
     </Stack>
